♻️ :: count 쿼리 -> size로 변경

+++ src/main/kotlin/com/example/noswimmingserver/domain/rank/domain/repository/CustomUserRankingRepository.kt
@@ -6,7 +6,7 @@ interface CustomUserRankingRepository {
 
     fun queryUserRanking(): List<UserRank>
 
-    fun countAllUsers(): Long
+    fun countAllUsers(): Int
 
-    fun countLessThanMe(userId: Long): Long
+    fun countLessThanMe(userId: Long): Int
 }
